Andreas G changes
Did some changes to bread_crumbs but duno if it is exactely what you
asked for.
modinfo changed date and author (to link up to forum)
I am going to push it to pamm for you this time but I've writen a basic
tutorial on how to do that yourself see in howto_pamm. If you have any
questions please ask me.
You should make a new map and push it soon so that you can make sure
you've got it.